rename extension folder to standards_export
